requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
attrs==24.3.0
blinker==1.9.0
cachelib==0.13.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
deprecation==2.1.0
Flask==3.1.0
Flask-Admin==1.6.1
Flask-Login==0.6.3
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.5.0
gotrue==2.11.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
msgspec==0.18.6
multidict==6.1.0
packaging==24.2
pillow==11.0.0
postgrest==0.18.0
propcache==0.2.1
psycopg2-binary==2.9.10
pydantic==2.10.4
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
qrcode==8.0
realtime==2.0.6
reportlab==4.2.5
requests==2.32.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
storage3==0.9.0
supabase==2.10.0
supafunc==0.7.0
typing_extensions==4.12.2
urllib3==2.2.3
websockets==13.1
Werkzeug==3.1.3
WTForms==3.2.1
yarl==1.18.3
